--- src/main/java/Commands/Commands.java
@@ -1,5 +1,0 @@
-package Commands;
-
-public enum Commands {
-    LIST, MARK, UNMARK, DELETE, BYE, TODO, EVENT, DEADLINE;
-}
