--- src/main/java/Commands/MarkTaskCommand.java
@@ -1,30 +1,0 @@
-package Commands;
-
-import Exceptions.DukeException;
-import Files.Storage;
-import Tasks.TaskList;
-import Ui.Ui;
-
-public class MarkTaskCommand extends Command {
-    private String taskIndex;
-
-    public MarkTaskCommand(String taskIndex) {
-        this.taskIndex = taskIndex;
-    }
-
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        try {
-            taskList.markTask(this.taskIndex);
-        } catch (DukeException e) {
-            System.out.println(e.getMessage());
-        }
-
-    }
-}
