--- src/main/java/Exceptions/DukeException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class DukeException extends RuntimeException {
-    public DukeException(String message, Throwable error) {
-        super(message, error);
-    }
-}
