--- src/main/java/Exceptions/IncompleteCommandException.java
@@ -1,9 +1,0 @@
-package Exceptions;
-
-import Exceptions.DukeException;
-
-public class IncompleteCommandException extends DukeException {
-    public IncompleteCommandException(String errorMessage, Throwable error) {
-        super(errorMessage,error);
-    }
-}
