--- src/main/java/Exceptions/InvalidIndexException.java
@@ -1,9 +1,0 @@
-package Exceptions;
-
-import Exceptions.DukeException;
-
-public class InvalidIndexException extends DukeException {
-    public InvalidIndexException(String errorMessage, Throwable error) {
-        super(errorMessage,error);
-    }
-}
