--- src/main/java/Exceptions/NoSuchTaskException.java
@@ -1,9 +1,0 @@
-package Exceptions;
-
-import Exceptions.DukeException;
-
-public class NoSuchTaskException extends DukeException {
-    public NoSuchTaskException(String errorMessage, Throwable error) {
-        super(errorMessage,error);
-    }
-}
