--- src/main/java/Exceptions/UnknownCommandException.java
@@ -1,9 +1,0 @@
-package Exceptions;
-
-import Exceptions.DukeException;
-
-public class UnknownCommandException extends DukeException {
-    public UnknownCommandException(String errorMessage, Throwable error) {
-        super(errorMessage,error);
-    }
-}
