--- src/main/java/Parsers/Parser.java
@@ -1,4 +1,0 @@
-package Parsers;
-
-public abstract class Parser {
-}
