+++ src/main/java/commands/Commands.java
@@ -1,0 +1,5 @@
+package commands;
+
+public enum Commands {
+    LIST, MARK, UNMARK, DELETE, BYE, TODO, EVENT, DEADLINE;
+}
