+++ src/main/java/exceptions/DukeException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class DukeException extends RuntimeException {
+    public DukeException(String message, Throwable error) {
+        super(message, error);
+    }
+}
