+++ src/main/java/exceptions/IncompleteCommandException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class IncompleteCommandException extends DukeException {
+    public IncompleteCommandException(String errorMessage, Throwable error) {
+        super(errorMessage,error);
+    }
+}
