+++ src/main/java/exceptions/InvalidIndexException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class InvalidIndexException extends DukeException {
+    public InvalidIndexException(String errorMessage, Throwable error) {
+        super(errorMessage,error);
+    }
+}
