+++ src/main/java/exceptions/NoSuchTaskException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class NoSuchTaskException extends DukeException {
+    public NoSuchTaskException(String errorMessage, Throwable error) {
+        super(errorMessage,error);
+    }
+}
