+++ src/main/java/exceptions/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class UnknownCommandException extends DukeException {
+    public UnknownCommandException(String errorMessage, Throwable error) {
+        super(errorMessage,error);
+    }
+}
