+++ src/main/java/parsers/Parser.java
@@ -1,0 +1,4 @@
+package parsers;
+
+public abstract class Parser {
+}
